+++ packages/db/src/domains/navigation/world/east-blue/island-registry.ts
@@ -1,0 +1,39 @@
+import { BARATIE } from './islands/baratie.js';
+import { CONOMI } from './islands/conomi.js';
+import { COZIA } from './islands/cozia.js';
+import { DAWN } from './islands/dawn.js';
+import { FRAUCE } from './islands/frauce.js';
+import { GECKO } from './islands/gecko.js';
+import { GOAT } from './islands/goat.js';
+import { KUMATE } from './islands/kumate.js';
+import { MIRRORBALL } from './islands/mirrorball.js';
+import { NAGAGUTSU } from './islands/nagagutsu.js';
+import { ORGAN } from './islands/organ.js';
+import { OYKOT } from './islands/oykot.js';
+import { POLE_STAR } from './islands/pole-star.js';
+import { RARE_ANIMAL } from './islands/rare-animal.js';
+import { SATSURUZO } from './islands/satsuruzo.js';
+import { SIXIS } from './islands/sixis.js';
+import { TEQUILA_WOLF } from './islands/tequila-wolf.js';
+import { YOTSUBA } from './islands/yotsuba.js';
+
+export const EAST_BLUE_ISLAND_REGISTRY = [
+  SATSURUZO,
+  DAWN,
+  GOAT,
+  YOTSUBA,
+  MIRRORBALL,
+  NAGAGUTSU,
+  ORGAN,
+  RARE_ANIMAL,
+  KUMATE,
+  SIXIS,
+  TEQUILA_WOLF,
+  GECKO,
+  BARATIE,
+  CONOMI,
+  COZIA,
+  FRAUCE,
+  OYKOT,
+  POLE_STAR,
+];
