+++ packages/db/src/domains/navigation/world/east-blue/islands/organ.ts
@@ -1,4 +1,4 @@
-import { defineIsland } from '../define-island.js';
+import { defineIsland } from '../../utils/define-island.js';
 
 export const ORGAN = defineIsland({
   key: 'organ',
