+++ packages/db/src/domains/navigation/world/east-blue/islands/yotsuba.ts
@@ -1,4 +1,4 @@
-import { defineIsland } from '../define-island.js';
+import { defineIsland } from '../../utils/define-island.js';
 
 export const YOTSUBA = defineIsland({
   key: 'yotsuba',
