+++ packages/db/src/domains/navigation/world/paradise/island-registry.ts
@@ -1,0 +1,8 @@
+import { ALABASTA } from './islands/alabasta.js';
+import { DRUM } from './islands/drum.js';
+import { LITTLE_GARDEN } from './islands/little-garden.js';
+import { REVERSE_MOUNTAIN } from './islands/reverse-mountain.js';
+import { WANO } from './islands/wano.js';
+import { WHISKY_PEAK } from './islands/whisky-peak.js';
+
+export const PARADISE_ISLAND_REGISTRY = [REVERSE_MOUNTAIN, WHISKY_PEAK, LITTLE_GARDEN, DRUM, ALABASTA, WANO];
