+++ packages/db/src/domains/navigation/world/paradise/islands/alabasta.ts
@@ -1,4 +1,4 @@
-import { defineIsland } from '../define-island.js';
+import { defineIsland } from '../../utils/define-island.js';
 
 export const ALABASTA = defineIsland({
   key: 'alabasta',
