+++ packages/db/src/domains/navigation/world/paradise/islands/little-garden.ts
@@ -1,4 +1,4 @@
-import { defineIsland } from '../define-island.js';
+import { defineIsland } from '../../utils/define-island.js';
 
 export const LITTLE_GARDEN = defineIsland({
   key: 'little_garden',
