+++ packages/db/src/domains/navigation/world/paradise/islands/reverse-mountain.ts
@@ -1,4 +1,4 @@
-import { defineIsland } from '../define-island.js';
+import { defineIsland } from '../../utils/define-island.js';
 
 export const REVERSE_MOUNTAIN = defineIsland({
   key: 'reverse_mountain',
