+++ packages/db/src/domains/navigation/world/paradise/islands/whisky-peak.ts
@@ -1,4 +1,4 @@
-import { defineIsland } from '../define-island.js';
+import { defineIsland } from '../../utils/define-island.js';
 
 export const WHISKY_PEAK = defineIsland({
   key: 'whisky_peak',
