--- packages/db/src/domains/navigation/world/paradise/registry.ts
@@ -1,8 +1,0 @@
-import { ALABASTA } from './alabasta.js';
-import { DRUM } from './drum.js';
-import { LITTLE_GARDEN } from './little-garden.js';
-import { REVERSE_MOUNTAIN } from './reverse-mountain.js';
-import { WANO } from './wano.js';
-import { WHISKY_PEAK } from './whisky-peak.js';
-
-export const PARADISE_ISLAND_REGISTRY = [REVERSE_MOUNTAIN, WHISKY_PEAK, LITTLE_GARDEN, DRUM, ALABASTA, WANO];
