+++ packages/db/src/domains/navigation/world/utils/define-island.ts
