+++ packages/db/src/domains/navigation/world/utils/in-buckets.ts
