+++ packages/db/src/domains/navigation/world/zones.ts
@@ -1,9 +1,8 @@
-import { ISLANDS, type Island } from './registry.js';
-import { SEAS, type Sea } from './seas.js';
+import { ISLANDS, ISLAND_LABELS } from './registry.js';
+import { SEAS, SEA_LABELS } from './seas.js';
 
 export const ZONES = [...ISLANDS, ...SEAS] as const;
 
 export type Zone = (typeof ZONES)[number];
 
-export { ISLANDS, type Island };
-export { SEAS, type Sea };
+export const ZONE_LABELS: Record<Zone, string> = { ...ISLAND_LABELS, ...SEA_LABELS };
